Move Rayo IQs to distinct files
