+++ src/main/java/org/jitsi/xmpp/extensions/rayo/RayoIqProvider.java
@@ -21,7 +21,6 @@ import org.jivesoftware.smack.packet.*;
 import org.jivesoftware.smack.parsing.*;
 import org.jivesoftware.smack.provider.*;
 import org.jivesoftware.smack.xml.*;
-import org.jxmpp.jid.Jid;
 
 import java.io.*;
 
@@ -34,7 +33,7 @@ import java.io.*;
  * @author Pawel Domas
  */
 public class RayoIqProvider
-    extends IqProvider<RayoIqProvider.RayoIq>
+    extends IqProvider<RayoIq>
 {
     /**
      * Rayo namespace.
@@ -220,358 +219,4 @@ public class RayoIqProvider
         return iq;
     }
 
-    /**
-     * Base class for all Ray IQs. Takes care of <header /> extension handling
-     * as well as other functions shared by all IQs.
-     */
-    public static abstract class RayoIq
-        extends IQ
-    {
-        /**
-         * Creates new instance of <tt>RayoIq</tt>.
-         *
-         * @param elementName the name of XML element that will be used.
-         */
-        protected RayoIq(String elementName)
-        {
-            super(elementName, NAMESPACE);
-        }
-
-        /**
-         * Creates new instance of this class as a copy from <tt>original</tt>.
-         *
-         * @param original the class to copy the data from.
-         */
-        protected RayoIq(RayoIq original)
-        {
-            super(original);
-        }
-
-        /**
-         * Returns value of the header extension with given <tt>name</tt>
-         * (if any).
-         * @param name the name of header extension which value we want to
-         *             retrieve.
-         * @return value of header extension with given <tt>name</tt> if it
-         *         exists or <tt>null</tt> otherwise.
-         */
-        public String getHeader(String name)
-        {
-            HeaderExtension header = findHeader(name);
-
-            return header != null ? header.getValue() : null;
-        }
-
-        private HeaderExtension findHeader(String name)
-        {
-            for (ExtensionElement ext: getExtensions())
-            {
-                if (ext instanceof HeaderExtension)
-                {
-                    HeaderExtension header = (HeaderExtension) ext;
-
-                    if (name.equals(header.getName()))
-                        return header;
-                }
-            }
-            return null;
-        }
-
-        /**
-         * Adds 'header' extension to this Rayo IQ with given name and value
-         * attributes.
-         * @param name the attribute name of the 'header' extension to be added.
-         * @param value the 'value' attribute of the 'header' extension that
-         *              will be added to this IQ.
-         */
-        public void setHeader(String name, String value)
-        {
-            HeaderExtension headerExt = findHeader(name);
-
-            if (headerExt == null)
-            {
-                headerExt = new HeaderExtension();
-
-                headerExt.setName(name);
-
-                addExtension(headerExt);
-            }
-
-            headerExt.setValue(value);
-        }
-    }
-
-    /**
-     * The 'dial' IQ used to initiate new outgoing call session in Rayo
-     * protocol.
-     */
-    public static class DialIq
-        extends RayoIq
-    {
-        /**
-         * The name of XML element for this IQ.
-         */
-        public static final String ELEMENT = "dial";
-
-        /**
-         * The name of source URI/address attribute. Referred as "source" to
-         * avoid confusion with "getFrom" and "setFrom" in {@link IQ} class.
-         */
-        public static final String SRC_ATTR_NAME = "from";
-
-        /**
-         * The name of destination URI/address attribute. Referred as "source"
-         * to avoid confusion with "getFrom" and "setFrom" in {@link IQ} class.
-         */
-        public static final String DST_ATTR_NAME = "to";
-
-        /**
-         * Source URI/address.
-         */
-        private String source;
-
-        /**
-         * Destination URI/address.
-         */
-        private String destination;
-
-        /**
-         * Creates new instance of <tt>DialIq</tt>.
-         */
-        public DialIq()
-        {
-            super(DialIq.ELEMENT);
-        }
-
-        /**
-         * Creates a new instance of this class as a copy from
-         * <tt>original</tt>.
-         * @param original the class to copy the data from.
-         */
-        public DialIq(DialIq original)
-        {
-            // copies: id, to, from, extensions, error, type
-            super(original);
-            source = original.source;
-            destination = original.destination;
-        }
-
-        /**
-         * Creates new <tt>DialIq</tt> for given source and destination
-         * addresses.
-         * @param to the destination address/call URI to be used.
-         * @param from the source address that will be set on
-         *             new <tt>DialIq</tt> instance.
-         * @return new <tt>DialIq</tt> parameterized with given source and
-         *         destination addresses.
-         */
-        public static DialIq create(String to, String from)
-        {
-            DialIq dialIq = new DialIq();
-
-            dialIq.setSource(from);
-
-            dialIq.setDestination(to);
-
-            return dialIq;
-        }
-
-        /**
-         * Return source address value set on this <tt>DialIq</tt>.
-         * @return source address value of this <tt>DialIq</tt>.
-         */
-        public String getSource()
-        {
-            return source;
-        }
-
-        /**
-         * Sets new source address value on this <tt>DialIq</tt>.
-         * @param source the new source address value to be set.
-         */
-        public void setSource(String source)
-        {
-            this.source = source;
-        }
-
-        /**
-         * Returns destination address/call URI associated with this instance.
-         * @return destination address/call URI associated with this instance.
-         */
-        public String getDestination()
-        {
-            return destination;
-        }
-
-        /**
-         * Sets new destination address/call URI on this <tt>DialIq</tt>.
-         * @param destination the new destination address/call URI to set.
-         */
-        public void setDestination(String destination)
-        {
-            this.destination = destination;
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        protected IQ.IQChildElementXmlStringBuilder getIQChildElementBuilder(
-            IQ.IQChildElementXmlStringBuilder xml)
-        {
-            xml.optAttribute(SRC_ATTR_NAME, source)
-                .optAttribute(DST_ATTR_NAME, destination);
-            xml.setEmptyElement();
-            return xml;
-        }
-    }
-
-    /**
-     * Rayo 'ref' IQ sent by the server as a reply to 'dial' request. Holds
-     * created call's resource in 'uri' attribute.
-     */
-    public static class RefIq
-        extends RayoIq
-    {
-        /**
-         * XML element name of <tt>RefIq</tt>.
-         */
-        public static final String ELEMENT = "ref";
-
-        /**
-         * Name of the URI attribute that stores call resource reference.
-         */
-        public static final String URI_ATTR_NAME = "uri";
-
-        /**
-         * Call resource/uri reference.
-         */
-        private String uri;
-
-        /**
-         * Creates new <tt>RefIq</tt>.
-         */
-        protected RefIq()
-        {
-            super(RefIq.ELEMENT);
-        }
-
-        /**
-         * Creates new <tt>RefIq</tt> parametrized with given call <tt>uri</tt>.
-         * @param uri the call URI to be set on newly created <tt>RefIq</tt>.
-         * @return new <tt>RefIq</tt> parametrized with given call <tt>uri</tt>.
-         */
-        public static RefIq create(String uri)
-        {
-            RefIq refIq = new RefIq();
-
-            refIq.setUri(uri);
-
-            return refIq;
-        }
-
-        /**
-         * Creates result <tt>RefIq</tt> for given <tt>requestIq</tt>
-         * parametrized with given call <tt>uri</tt>.
-         * @param requestIq the request IQ which 'from', 'to' and 'id'
-         *                  attributes will be used for constructing result IQ.
-         * @param uri the call URI that will be included in newly created
-         *            <tt>RefIq</tt>.
-         * @return result <tt>RefIq</tt> for given <tt>requestIq</tt>
-         *         parametrized with given call <tt>uri</tt>.
-         */
-        public static RefIq createResult(IQ requestIq, String uri)
-        {
-            RefIq refIq = create(uri);
-
-            refIq.setType(IQ.Type.result);
-            refIq.setStanzaId(requestIq.getStanzaId());
-            refIq.setFrom(requestIq.getTo());
-            refIq.setTo(requestIq.getFrom());
-
-            return refIq;
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        protected IQ.IQChildElementXmlStringBuilder getIQChildElementBuilder(
-            IQ.IQChildElementXmlStringBuilder xml)
-        {
-            xml.optAttribute(URI_ATTR_NAME, uri);
-            xml.setEmptyElement();
-            return xml;
-        }
-
-        /**
-         * Sets given call <tt>uri</tt> value on this instance.
-         * @param uri the call <tt>uri</tt> to be stored in this instance.
-         */
-        public void setUri(String uri)
-        {
-            this.uri = uri;
-        }
-
-        /**
-         * Returns call URI held by this instance.
-         * @return the call URI held by this instance.
-         */
-        public String getUri()
-        {
-            return uri;
-        }
-    }
-
-    /**
-     * Rayo hangup IQ is sent by the controlling agent to tell the server that
-     * call whose resource is mentioned in IQ's 'to' attribute should be
-     * terminated. Server immediately replies with result IQ which means that
-     * hangup operation is now scheduled. After it is actually executed presence
-     * indication with {@link EndExtension} is sent through the presence to
-     * confirm the operation.
-     */
-    public static class HangUp
-        extends RayoIq
-    {
-        /**
-         * The name of 'hangup' element.
-         */
-        public static final String ELEMENT = "hangup";
-
-        /**
-         * Creates new instance of <tt>HangUp</tt> IQ.
-         */
-        protected HangUp()
-        {
-            super(ELEMENT);
-        }
-
-        /**
-         * Creates new, parametrized instance of {@link HangUp} IQ.
-         * @param from source JID.
-         * @param to the destination address/call URI to be ended by this IQ.
-         * @return new, parametrized instance of {@link HangUp} IQ.
-         */
-        public static HangUp create(Jid from, Jid to)
-        {
-            HangUp hangUp = new HangUp();
-            hangUp.setFrom(from);
-            hangUp.setTo(to);
-            hangUp.setType(Type.set);
-
-            return hangUp;
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        protected IQ.IQChildElementXmlStringBuilder getIQChildElementBuilder(
-            IQ.IQChildElementXmlStringBuilder xml)
-        {
-            xml.setEmptyElement();
-            return xml;
-        }
-    }
 }
